requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database (SQLAlchemy + async)
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# HTTP client
httpx>=0.25.0

# Blockchain libraries
web3>=6.11.0
solders>=0.18.0
solana>=0.30.2

# Rate limiting
slowapi>=0.1.9

# Caching
redis>=5.0.0
hiredis>=2.2.3

# Monitoring & Observability
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# Background tasks
celery>=5.3.4
flower>=2.0.1

# Security
cryptography>=41.0.7
pyjwt>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Development tools
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.0

# Production server
gunicorn>=21.2.0
